ways to earn complete

+++ db/migrate/20221108131308_create_ways_to_earn_points.rb
@@ -2,7 +2,7 @@ class CreateWaysToEarnPoints < ActiveRecord::Migration[6.0]
   def change
     create_table :ways_to_earn_points do |t|
       t.string :title
-      t.integer :points_awarded
+      t.float :points_awarded
       t.boolean :status
       t.text :summary
       t.references :shop, null: false, foreign_key: true
